feat: Milestone 0 inventory + Milestone 1 routes/apps.py blueprint
Agent-Logs-Url: https://github.com/abartrim/sobs/sessions/30b0c283-47d6-4ae8-9f23-779024f9e70b
