+++ app.py
@@ -10295,227 +10295,6 @@ async def ingest_errors():
     return jsonify({"ok": True}), 200
 
 
-# ---------------------------------------------------------------------------
-# App / Release / Artifact Registry APIs (Phase 1 scaffolding)
-# ---------------------------------------------------------------------------
-@app.route("/v1/apps", methods=["GET"])
-@require_api_key
-async def list_apps():
-    db = get_db()
-    q = (request.args.get("q") or "").strip().lower()
-    rows = [dict(r) for r in db.execute("SELECT * FROM sobs_apps FINAL WHERE IsDeleted=0 ORDER BY Name ASC").fetchall()]
-    apps = [_serialize_app_row(row) for row in rows]
-    if q:
-        apps = [item for item in apps if q in item["name"].lower() or q in item["slug"].lower()]
-    return jsonify(apps), 200
-
-
-@app.route("/v1/apps", methods=["POST"])
-@require_api_key
-async def create_app_registry_entry():
-    db = get_db()
-    payload = await request.get_json(force=True, silent=True) or {}
-    name = str(payload.get("name", "")).strip()
-    if not name:
-        return jsonify({"error": "name is required"}), 400
-
-    slug = _app_slug(str(payload.get("slug", "")).strip() or name)
-    existing = db.execute(
-        "SELECT Id FROM sobs_apps FINAL WHERE Slug=? AND IsDeleted=0 LIMIT 1",
-        [slug],
-    ).fetchone()
-    if existing:
-        return jsonify({"error": "app slug already exists"}), 409
-
-    version = int(time.time() * 1000)
-    app_id = str(payload.get("id", "")).strip() or uuid.uuid4().hex
-    row = {
-        "Id": app_id,
-        "Name": name,
-        "Slug": slug,
-        "OwnerTeam": str(payload.get("ownerTeam", "")).strip(),
-        "RepoUrl": str(payload.get("repoUrl", "")).strip(),
-        "DefaultEnvironment": str(payload.get("defaultEnvironment", "")).strip(),
-        "Enabled": 1 if _parse_bool(payload.get("enabled", True), True) else 0,
-        "MetadataJson": _safe_json_dumps(payload.get("metadata", {})),
-        "IsDeleted": 0,
-        "Version": version,
-        "CreatedAt": _now_iso(),
-        "UpdatedAt": _now_iso(),
-    }
-    _insert_rows_json_each_row(db, "sobs_apps", [row])
-    return jsonify(_serialize_app_row(row)), 201
-
-
-@app.route("/v1/apps/<app_id>", methods=["GET"])
-@require_api_key
-async def get_app_registry_entry(app_id: str):
-    db = get_db()
-    row = _find_app_by_id(db, app_id)
-    if not row:
-        return jsonify({"error": "not found"}), 404
-    return jsonify(_serialize_app_row(row)), 200
-
-
-@app.route("/v1/apps/<app_id>", methods=["PATCH"])
-@require_api_key
-async def update_app_registry_entry(app_id: str):
-    db = get_db()
-    current = _find_app_by_id(db, app_id)
-    if not current:
-        return jsonify({"error": "not found"}), 404
-
-    payload = await request.get_json(force=True, silent=True) or {}
-    name = str(payload.get("name", current.get("Name", ""))).strip()
-    if not name:
-        return jsonify({"error": "name is required"}), 400
-
-    slug = _app_slug(str(payload.get("slug", current.get("Slug", ""))).strip() or name)
-    conflict = db.execute(
-        "SELECT Id FROM sobs_apps FINAL WHERE Slug=? AND IsDeleted=0 AND Id!=? LIMIT 1",
-        [slug, app_id],
-    ).fetchone()
-    if conflict:
-        return jsonify({"error": "app slug already exists"}), 409
-
-    version = int(time.time() * 1000)
-    row = {
-        "Id": app_id,
-        "Name": name,
-        "Slug": slug,
-        "OwnerTeam": str(payload.get("ownerTeam", current.get("OwnerTeam", ""))).strip(),
-        "RepoUrl": str(payload.get("repoUrl", current.get("RepoUrl", ""))).strip(),
-        "DefaultEnvironment": str(payload.get("defaultEnvironment", current.get("DefaultEnvironment", ""))).strip(),
-        "Enabled": 1 if _parse_bool(payload.get("enabled", int(current.get("Enabled", 1))), True) else 0,
-        "MetadataJson": _safe_json_dumps(
-            payload.get("metadata", _safe_json_loads(current.get("MetadataJson", ""), {}))
-        ),
-        "IsDeleted": 0,
-        "Version": version,
-        "CreatedAt": str(current.get("CreatedAt", "")) or _now_iso(),
-        "UpdatedAt": _now_iso(),
-    }
-    _insert_rows_json_each_row(db, "sobs_apps", [row])
-    return jsonify(_serialize_app_row(row)), 200
-
-
-@app.route("/v1/apps/<app_id>/releases", methods=["GET"])
-@require_api_key
-async def list_app_releases(app_id: str):
-    db = get_db()
-    app_row = _find_app_by_id(db, app_id)
-    if not app_row:
-        return jsonify({"error": "app not found"}), 404
-    rows = [
-        _serialize_release_row(dict(r))
-        for r in db.execute(
-            "SELECT * FROM sobs_app_releases FINAL WHERE AppId=? AND IsDeleted=0 ORDER BY ReleasedAt DESC",
-            [app_id],
-        ).fetchall()
-    ]
-    return jsonify(rows), 200
-
-
-@app.route("/v1/apps/<app_id>/releases", methods=["POST"])
-@require_api_key
-async def create_app_release(app_id: str):
-    db = get_db()
-    app_row = _find_app_by_id(db, app_id)
-    if not app_row:
-        return jsonify({"error": "app not found"}), 404
-
-    payload = await request.get_json(force=True, silent=True) or {}
-    release_version = str(payload.get("version", "")).strip()
-    if not release_version:
-        return jsonify({"error": "version is required"}), 400
-
-    version = int(time.time() * 1000)
-    row = {
-        "Id": str(payload.get("id", "")).strip() or uuid.uuid4().hex,
-        "AppId": app_id,
-        "ReleaseVersion": release_version,
-        "CommitSha": str(payload.get("commitSha", "")).strip(),
-        "BuildId": str(payload.get("buildId", "")).strip(),
-        "Environment": str(payload.get("environment", "")).strip(),
-        "ReleasedAt": str(payload.get("releasedAt", "")).strip() or _now_iso(),
-        "MetadataJson": _safe_json_dumps(payload.get("metadata", {})),
-        "IsDeleted": 0,
-        "Version": version,
-    }
-    _insert_rows_json_each_row(db, "sobs_app_releases", [row])
-    return jsonify(_serialize_release_row(row)), 201
-
-
-@app.route("/v1/releases/<release_id>", methods=["GET"])
-@require_api_key
-async def get_release(release_id: str):
-    db = get_db()
-    row = _find_release_by_id(db, release_id)
-    if not row:
-        return jsonify({"error": "not found"}), 404
-
-    release = _serialize_release_row(row)
-    artifacts = [
-        _serialize_artifact_row(dict(r))
-        for r in db.execute(
-            "SELECT * FROM sobs_release_artifacts FINAL WHERE ReleaseId=? AND IsDeleted=0 ORDER BY UploadedAt DESC",
-            [release_id],
-        ).fetchall()
-    ]
-    return jsonify({"release": release, "artifacts": artifacts}), 200
-
-
-@app.route("/v1/releases/<release_id>/artifacts", methods=["GET"])
-@require_api_key
-async def list_release_artifacts(release_id: str):
-    db = get_db()
-    row = _find_release_by_id(db, release_id)
-    if not row:
-        return jsonify({"error": "release not found"}), 404
-    artifacts = [
-        _serialize_artifact_row(dict(r))
-        for r in db.execute(
-            "SELECT * FROM sobs_release_artifacts FINAL WHERE ReleaseId=? AND IsDeleted=0 ORDER BY UploadedAt DESC",
-            [release_id],
-        ).fetchall()
-    ]
-    return jsonify(artifacts), 200
-
-
-@app.route("/v1/releases/<release_id>/artifacts/meta", methods=["POST"])
-@require_api_key
-async def create_release_artifact_meta(release_id: str):
-    db = get_db()
-    release = _find_release_by_id(db, release_id)
-    if not release:
-        return jsonify({"error": "release not found"}), 404
-
-    payload = await request.get_json(force=True, silent=True) or {}
-    artifact_type = str(payload.get("artifactType", "")).strip()
-    name = str(payload.get("name", "")).strip()
-    if not artifact_type or not name:
-        return jsonify({"error": "artifactType and name are required"}), 400
-
-    version = int(time.time() * 1000)
-    row = {
-        "Id": str(payload.get("id", "")).strip() or uuid.uuid4().hex,
-        "ReleaseId": release_id,
-        "ArtifactType": artifact_type,
-        "Name": name,
-        "ContentType": str(payload.get("contentType", "")).strip(),
-        "Size": int(payload.get("size", 0) or 0),
-        "StorageRef": str(payload.get("storageRef", "")).strip(),
-        "ChecksumSha256": str(payload.get("checksumSha256", "")).strip(),
-        "Platform": str(payload.get("platform", "")).strip(),
-        "Architecture": str(payload.get("architecture", "")).strip(),
-        "MetadataJson": _safe_json_dumps(payload.get("metadata", {})),
-        "UploadedAt": str(payload.get("uploadedAt", "")).strip() or _now_iso(),
-        "IsDeleted": 0,
-        "Version": version,
-    }
-    _insert_rows_json_each_row(db, "sobs_release_artifacts", [row])
-    return jsonify(_serialize_artifact_row(row)), 201
-
 
 ERROR_SOURCES_SQL = """
 SELECT
@@ -33928,6 +33707,15 @@ async def api_onboarding_create_issues():
                 )
 
     return jsonify(results)
+
+
+# ---------------------------------------------------------------------------
+# Blueprint registration (deferred – placed here so all app.py helpers are
+# already defined when the blueprint modules import from app).
+# ---------------------------------------------------------------------------
+from routes.apps import apps_bp as _apps_bp  # noqa: E402  # type: ignore[has-type]
+
+app.register_blueprint(_apps_bp)  # type: ignore[has-type]
 
 
 if __name__ == "__main__":
